+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## v1.4.0 - 2021-12-12
+
+- Handle RSS and Atom as formats in addition to Bookmarks Json format
+
 ## v1.3.2 - 2020-05-08
-Try to fix an issue where the sync would stop working after a while.
+
+- Try to fix an issue where the sync would stop working after a while.